Replace golib/app with app2 pkg
