+++ apps/golib/app/config.go
@@ -1,4 +1,4 @@
-package app2
+package app
 
 import (
 	"context"
