+++ apps/golib/app/context.go
@@ -3,18 +3,27 @@ package app
 import (
 	"context"
 
-	"github.com/kneadCODE/crazycat/apps/golib/app/config"
-	"github.com/newrelic/go-agent/v3/newrelic"
+	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
 )
 
 // ConfigFromContext retrieves the Config from context if exists else return a new Config
-func ConfigFromContext(ctx context.Context) config.Config {
-	if v, ok := ctx.Value(configCtxKey).(config.Config); ok {
+func ConfigFromContext(ctx context.Context) Config {
+	if v, ok := ctx.Value(configCtxKey).(Config); ok {
 		return v
 	}
-	return config.Config{}
+	return Config{}
+}
+
+// ContextWithAttributes adds the given attributes to the context and the span in the context (if any)
+func ContextWithAttributes(ctx context.Context, attrs ...attribute.KeyValue) context.Context {
+	span := trace.SpanFromContext(ctx)
+	span.SetAttributes(attrs...)
+
+	ctx = setOTELAttrsInContext(ctx, append(otelAttrsFromContext(ctx), attrs...))
+
+	return ctx
 }
 
 // contextKey implementation is referenced from go stdlib:
@@ -26,45 +35,34 @@ type contextKey struct {
 func (k contextKey) String() string { return "app context value " + k.name }
 
 var (
-	configCtxKey     = contextKey{"app-config"}
-	zapCtxKey        = contextKey{"app-zap"}
-	otelTracerCtxKey = contextKey{"app-otel-tracer"}
-	newrelicCtxKey   = contextKey{"app_newrelic"}
+	configCtxKey    = contextKey{"app-config"}
+	zapCtxKey       = contextKey{"app-zap"}
+	otelAttrsCtxKey = contextKey{"app-otel-attrs"}
+	// newrelicCtxKey   = contextKey{"app_newrelic"}
 )
 
-func setConfigInContext(ctx context.Context, cfg config.Config) context.Context {
+func setConfigInContext(ctx context.Context, cfg Config) context.Context {
 	return context.WithValue(ctx, configCtxKey, cfg)
 }
 
-func setZapInContext(ctx context.Context, l *zap.SugaredLogger) context.Context {
+func setZapInContext(ctx context.Context, l *zap.Logger) context.Context {
 	return context.WithValue(ctx, zapCtxKey, l)
 }
 
-func zapFromContext(ctx context.Context) *zap.SugaredLogger {
-	if l, ok := ctx.Value(zapCtxKey).(*zap.SugaredLogger); ok {
-		return l
+func zapFromContext(ctx context.Context) *zap.Logger {
+	if v, ok := ctx.Value(zapCtxKey).(*zap.Logger); ok {
+		return v
 	}
 	return nil
 }
 
-func setOTELTracerInContext(ctx context.Context, tracer trace.Tracer) context.Context {
-	return context.WithValue(ctx, otelTracerCtxKey, tracer)
+func setOTELAttrsInContext(ctx context.Context, attrs []attribute.KeyValue) context.Context {
+	return context.WithValue(ctx, otelAttrsCtxKey, attrs)
 }
 
-func otelTracerFromContext(ctx context.Context) trace.Tracer {
-	if tracer, ok := ctx.Value(otelTracerCtxKey).(trace.Tracer); ok {
-		return tracer
-	}
-	return nil
-}
-
-func setNewRelicInContext(ctx context.Context, nrApp *newrelic.Application) context.Context {
-	return context.WithValue(ctx, newrelicCtxKey, nrApp)
-}
-
-func newRelicFromContext(ctx context.Context) *newrelic.Application {
-	if nrApp, ok := ctx.Value(newrelicCtxKey).(*newrelic.Application); ok {
-		return nrApp
+func otelAttrsFromContext(ctx context.Context) []attribute.KeyValue {
+	if v, ok := ctx.Value(otelAttrsCtxKey).([]attribute.KeyValue); ok {
+		return v
 	}
 	return nil
 }
