+++ apps/golib/app/init_test.go
@@ -3,392 +3,192 @@ package app
 import (
 	"context"
 	"errors"
-	"fmt"
-	"os"
 	"testing"
 
-	"github.com/getsentry/sentry-go"
-	"github.com/kneadCODE/crazycat/apps/golib/app/config"
-	"github.com/kneadCODE/crazycat/apps/golib/app/internal"
-	"github.com/newrelic/go-agent/v3/newrelic"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/propagation"
+	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/resource"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
+	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
 )
 
 func TestInit(t *testing.T) {
 	type testCase struct {
-		givenConfig        config.Config
-		givenConfigErr     error
-		givenZap           *zap.SugaredLogger
-		givenZapErr        error
-		givenSentry        *sentry.Hub
-		givenSentryErr     error
-		givenNR            *newrelic.Application
-		givenNRErr         error
-		givenOTELShutdownF func(context.Context) error
-		givenOTELTracerErr error
-		expErr             error
-		expCfg             config.Config
-		expSentry          bool
-		expNR              bool
+		givenSentryEnabled                    bool
+		mockRes                               *resource.Resource
+		mockResErr                            error
+		mockDebugMode                         bool
+		mockPropagators                       propagation.TextMapPropagator
+		mockZap                               *zap.Logger
+		mockZapErr                            error
+		mockTraceProv                         *sdktrace.TracerProvider
+		mockTraceProvErr                      error
+		mockMeterProv                         *sdkmetric.MeterProvider
+		mockMeterProvErr                      error
+		expCfg                                Config
+		expErr                                error
+		expNewOTELResourceFromEnvStubCalled   bool
+		expNewZapStubCalled                   bool
+		expNewOTELPropagatorStubCalled        bool
+		expNewOTELTraceProviderStubCalled     bool
+		expNewOTELMeterProviderStubCalled     bool
+		expSetOTELTextMapPropagatorStubCalled bool
+		expSetOTELTracerProviderStubCalled    bool
+		expSetOTELMeterProviderStubCalled     bool
 	}
 
 	tcs := map[string]testCase{
-		"ok: basic setup, prod env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:           zap.NewExample().Sugar(),
-			givenOTELShutdownF: func(context.Context) error { return nil },
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
+		"res err": {
+			mockResErr:                          errors.New("some err"),
+			expErr:                              errors.New("some err"),
+			expNewOTELResourceFromEnvStubCalled: true,
 		},
-		"ok: basic setup, staging env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "staging",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:           zap.NewExample().Sugar(),
-			givenOTELShutdownF: func(context.Context) error { return nil },
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "staging",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
+		"invalid env": {
+			mockRes:                             resource.NewWithAttributes(semconv.SchemaURL, semconv.DeploymentEnvironment("dev")),
+			expErr:                              errors.New("invalid env: [dev]"),
+			expNewOTELResourceFromEnvStubCalled: true,
 		},
-		"ok: basic setup, dev env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "development",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:           zap.NewExample().Sugar(),
-			givenOTELShutdownF: func(context.Context) error { return nil },
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "development",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
+		"zap err": {
+			mockRes:                               resource.NewWithAttributes(semconv.SchemaURL, semconv.DeploymentEnvironment("development")),
+			mockDebugMode:                         true,
+			mockPropagators:                       propagation.NewCompositeTextMapPropagator(),
+			mockZapErr:                            errors.New("some err"),
+			expErr:                                errors.New("some err"),
+			expNewOTELResourceFromEnvStubCalled:   true,
+			expNewOTELPropagatorStubCalled:        true,
+			expSetOTELTextMapPropagatorStubCalled: true,
+			expNewZapStubCalled:                   true,
 		},
-		"ok: with sentry, prod env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:           zap.NewExample().Sugar(),
-			givenSentry:        sentry.NewHub(nil, nil),
-			givenOTELShutdownF: func(context.Context) error { return nil },
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			expSentry: true,
+		"trace provider err": {
+			mockRes:                               resource.NewWithAttributes(semconv.SchemaURL, semconv.DeploymentEnvironment("development")),
+			mockDebugMode:                         true,
+			mockPropagators:                       propagation.NewCompositeTextMapPropagator(),
+			mockZap:                               zap.NewExample(),
+			mockTraceProvErr:                      errors.New("some err"),
+			expErr:                                errors.New("some err"),
+			expNewOTELResourceFromEnvStubCalled:   true,
+			expNewOTELPropagatorStubCalled:        true,
+			expSetOTELTextMapPropagatorStubCalled: true,
+			expNewZapStubCalled:                   true,
+			expNewOTELTraceProviderStubCalled:     true,
 		},
-		"err: with cfg error, prod env": {
-			givenConfigErr: errors.New("some cfg err"),
-			expErr:         errors.New("some cfg err"),
+		"meter provider err": {
+			mockRes:                               resource.NewWithAttributes(semconv.SchemaURL, semconv.DeploymentEnvironment("development")),
+			mockDebugMode:                         true,
+			mockPropagators:                       propagation.NewCompositeTextMapPropagator(),
+			mockZap:                               zap.NewExample(),
+			mockTraceProv:                         sdktrace.NewTracerProvider(),
+			mockMeterProvErr:                      errors.New("some err"),
+			expErr:                                errors.New("some err"),
+			expNewOTELResourceFromEnvStubCalled:   true,
+			expNewOTELPropagatorStubCalled:        true,
+			expSetOTELTextMapPropagatorStubCalled: true,
+			expNewZapStubCalled:                   true,
+			expNewOTELTraceProviderStubCalled:     true,
+			expSetOTELTracerProviderStubCalled:    true,
+			expNewOTELMeterProviderStubCalled:     true,
 		},
-		"err: with zap error, prod env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZapErr: errors.New("some zap err"),
-			expErr:      errors.New("some zap err"),
-		},
-		"err: with sentry error, prod env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:       zap.NewExample().Sugar(),
-			givenSentryErr: errors.New("some sentry err"),
-			expErr:         errors.New("some sentry err"),
-		},
-		"ok: with NR, prod env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:           zap.NewExample().Sugar(),
-			givenNR:            &newrelic.Application{},
-			givenOTELShutdownF: func(context.Context) error { return nil },
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			expNR: true,
-		},
-		"err: with NR error, prod env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:   zap.NewExample().Sugar(),
-			givenNRErr: errors.New("some NR err"),
-			expErr:     errors.New("some NR err"),
-		},
-		"err: with OTEL error, prod env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:           zap.NewExample().Sugar(),
-			givenOTELTracerErr: errors.New("some OTEL err"),
-			expErr:             errors.New("some OTEL err"),
-		},
-		"ok: with everything, prod env": {
-			givenConfig: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			givenZap:           zap.NewExample().Sugar(),
-			givenSentry:        sentry.NewHub(nil, nil),
-			givenNR:            &newrelic.Application{},
-			givenOTELShutdownF: func(context.Context) error { return nil },
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-			expSentry: true,
-			expNR:     true,
+		"success": {
+			mockRes:                               resource.NewWithAttributes(semconv.SchemaURL, semconv.DeploymentEnvironment("development")),
+			mockDebugMode:                         true,
+			mockPropagators:                       propagation.NewCompositeTextMapPropagator(),
+			mockZap:                               zap.NewExample(),
+			mockTraceProv:                         sdktrace.NewTracerProvider(),
+			mockMeterProv:                         sdkmetric.NewMeterProvider(),
+			expCfg:                                Config{Env: EnvDev, res: resource.NewWithAttributes(semconv.SchemaURL, semconv.DeploymentEnvironment("development"))},
+			expNewOTELResourceFromEnvStubCalled:   true,
+			expNewOTELPropagatorStubCalled:        true,
+			expSetOTELTextMapPropagatorStubCalled: true,
+			expNewZapStubCalled:                   true,
+			expNewOTELTraceProviderStubCalled:     true,
+			expSetOTELTracerProviderStubCalled:    true,
+			expNewOTELMeterProviderStubCalled:     true,
+			expSetOTELMeterProviderStubCalled:     true,
 		},
 	}
 
 	for name, tc := range tcs {
 		t.Run(name, func(t *testing.T) {
 			// Given:
-			defer func() {
-				newConfigFromEnvF = newConfigFromEnv
-				newZapF = internal.NewZap
-				newSentryF = internal.NewSentryHub
-				newNewRelicF = internal.NewNewRelicApp
-				initOTELF = internal.InitOTEL
-			}()
-			newConfigFromEnvF = func() (config.Config, error) {
-				return tc.givenConfig, tc.givenConfigErr
+			defer resetStubs()
+			var newOTELResourceFromEnvStubCalled bool
+			newOTELResourceFromEnvStub = func(ctx context.Context) (*resource.Resource, error) {
+				newOTELResourceFromEnvStubCalled = true
+				return tc.mockRes, tc.mockResErr
 			}
-			newZapF = func(cfg config.Config) (*zap.SugaredLogger, error) {
-				require.Equal(t, tc.givenConfig, cfg)
-				return tc.givenZap, tc.givenZapErr
+			var newZapStubCalled bool
+			newZapStub = func(debugMode bool, res *resource.Resource) (*zap.Logger, error) {
+				newZapStubCalled = true
+				require.Equal(t, tc.mockRes, res)
+				require.Equal(t, tc.mockDebugMode, debugMode)
+				return tc.mockZap, tc.mockZapErr
 			}
-			newSentryF = func(cfg config.Config) (*sentry.Hub, error) {
-				require.Equal(t, tc.givenConfig, cfg)
-				return tc.givenSentry, tc.givenSentryErr
+			var newOTELPropagatorStubCalled bool
+			newOTELPropagatorStub = func(isSentryEnabled bool) propagation.TextMapPropagator {
+				newOTELPropagatorStubCalled = true
+				require.Equal(t, tc.givenSentryEnabled, isSentryEnabled)
+				return tc.mockPropagators
 			}
-			newNewRelicF = func(cfg config.Config) (*newrelic.Application, error) {
-				require.Equal(t, tc.givenConfig, cfg)
-				return tc.givenNR, tc.givenNRErr
+			var newOTELTraceProviderStubCalled bool
+			newOTELTraceProviderStub = func(res *resource.Resource, isSentryEnabled bool) (*sdktrace.TracerProvider, error) {
+				newOTELTraceProviderStubCalled = true
+				require.Equal(t, tc.givenSentryEnabled, isSentryEnabled)
+				require.Equal(t, tc.mockRes, res)
+				return tc.mockTraceProv, tc.mockTraceProvErr
 			}
-			initOTELF = func(ctx context.Context, cfg config.Config, isSentryEnabled bool) (func(context.Context) error, error) {
-				require.Equal(t, tc.givenConfig, cfg)
-				// TODO: Validate isSentryEnabled's impact somehow
-				return tc.givenOTELShutdownF, tc.givenOTELTracerErr
+			var newOTELMeterProviderStubCalled bool
+			newOTELMeterProviderStub = func(res *resource.Resource) (*sdkmetric.MeterProvider, error) {
+				newOTELMeterProviderStubCalled = true
+				require.Equal(t, tc.mockRes, res)
+				return tc.mockMeterProv, tc.mockMeterProvErr
+			}
+			var setOTELTextMapPropagatorStubCalled bool
+			setOTELTextMapPropagatorStub = func(propagator propagation.TextMapPropagator) {
+				setOTELTextMapPropagatorStubCalled = true
+				require.Equal(t, tc.mockPropagators, propagator)
+			}
+			var setOTELTracerProviderStubCalled bool
+			setOTELTracerProviderStub = func(tp trace.TracerProvider) {
+				setOTELTracerProviderStubCalled = true
+				require.Equal(t, tc.mockTraceProv, tp)
+			}
+			var setOTELMeterProviderStubCalled bool
+			setOTELMeterProviderStub = func(mp metric.MeterProvider) {
+				setOTELMeterProviderStubCalled = true
+				require.Equal(t, tc.mockMeterProv, mp)
 			}
 
 			// When:
 			ctx, finish, err := Init()
 
 			// Then:
+			require.Equal(t, tc.expNewOTELResourceFromEnvStubCalled, newOTELResourceFromEnvStubCalled)
+			require.Equal(t, tc.expNewZapStubCalled, newZapStubCalled)
+			require.Equal(t, tc.expNewOTELPropagatorStubCalled, newOTELPropagatorStubCalled)
+			require.Equal(t, tc.expNewOTELTraceProviderStubCalled, newOTELTraceProviderStubCalled)
+			require.Equal(t, tc.expNewOTELMeterProviderStubCalled, newOTELMeterProviderStubCalled)
+			require.Equal(t, tc.expSetOTELTextMapPropagatorStubCalled, setOTELTextMapPropagatorStubCalled)
+			require.Equal(t, tc.expSetOTELTracerProviderStubCalled, setOTELTracerProviderStubCalled)
+			require.Equal(t, tc.expSetOTELMeterProviderStubCalled, setOTELMeterProviderStubCalled)
+
 			if tc.expErr != nil {
-				require.EqualError(t, err, tc.expErr.Error())
+				require.Equal(t, tc.expErr, err)
 				require.Nil(t, finish)
+				require.EqualValues(t, tc.expCfg, ConfigFromContext(ctx))
 			} else {
 				require.Nil(t, err)
 				require.NotNil(t, finish)
 
 				cfg := ConfigFromContext(ctx)
 				require.EqualValues(t, tc.expCfg, cfg)
-
-				require.Equal(t, tc.givenZap, zapFromContext(ctx))
-				require.Equal(t, tc.givenSentry, sentry.GetHubFromContext(ctx))
-				require.Equal(t, tc.givenNR, newRelicFromContext(ctx))
+				require.Equal(t, tc.mockZap, zapFromContext(ctx))
 
 				finish()
 			}
-		})
-	}
-}
-
-func Test_NewConfigFromEnv(t *testing.T) {
-	type testCase struct {
-		givenOSEnvF func(t *testing.T)
-		expErr      error
-		expCfg      config.Config
-	}
-
-	tcs := map[string]testCase{
-		"ok: prod env": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNameKey), "name"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceNamespaceKey), "project"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceVersionKey), "v1.0.0"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceInstanceIDKey), "instance"))
-				require.NoError(t, os.Setenv(string(semconv.DeploymentEnvironmentKey), "production"))
-			},
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "production",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-		},
-		"ok: staging env": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNameKey), "name"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceNamespaceKey), "project"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceVersionKey), "v1.0.0"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceInstanceIDKey), "instance"))
-				require.NoError(t, os.Setenv(string(semconv.DeploymentEnvironmentKey), "staging"))
-			},
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "staging",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-		},
-		"ok: dev env": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNameKey), "name"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceNamespaceKey), "project"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceVersionKey), "v1.0.0"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceInstanceIDKey), "instance"))
-				require.NoError(t, os.Setenv(string(semconv.DeploymentEnvironmentKey), "development"))
-			},
-			expCfg: config.Config{
-				Name:             "name",
-				Env:              "development",
-				Project:          "project",
-				Version:          "v1.0.0",
-				ServerInstanceID: "instance",
-			},
-		},
-		"err: empty name": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNamespaceKey), "project"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceVersionKey), "v1.0.0"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceInstanceIDKey), "instance"))
-				require.NoError(t, os.Setenv(string(semconv.DeploymentEnvironmentKey), "production"))
-			},
-			expErr: fmt.Errorf("name is empty: %w", config.ErrInvalidConfig),
-		},
-		"err: empty project": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNameKey), "name"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceVersionKey), "v1.0.0"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceInstanceIDKey), "instance"))
-				require.NoError(t, os.Setenv(string(semconv.DeploymentEnvironmentKey), "production"))
-			},
-			expErr: fmt.Errorf("project is empty: %w", config.ErrInvalidConfig),
-		},
-		"err: empty version": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNameKey), "name"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceNamespaceKey), "project"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceInstanceIDKey), "instance"))
-				require.NoError(t, os.Setenv(string(semconv.DeploymentEnvironmentKey), "production"))
-			},
-			expErr: fmt.Errorf("version is empty: %w", config.ErrInvalidConfig),
-		},
-		"err: empty server instance id": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNameKey), "name"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceNamespaceKey), "project"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceVersionKey), "v1.0.0"))
-				require.NoError(t, os.Setenv(string(semconv.DeploymentEnvironmentKey), "production"))
-			},
-			expErr: fmt.Errorf("server instance ID is empty: %w", config.ErrInvalidConfig),
-		},
-		"err: empty env": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNameKey), "name"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceNamespaceKey), "project"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceVersionKey), "v1.0.0"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceInstanceIDKey), "instance"))
-			},
-			expErr: fmt.Errorf("invalid env: %w", config.ErrInvalidConfig),
-		},
-		"err: invalid env": {
-			givenOSEnvF: func(t *testing.T) {
-				require.NoError(t, os.Setenv(string(semconv.ServiceNameKey), "name"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceNamespaceKey), "project"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceVersionKey), "v1.0.0"))
-				require.NoError(t, os.Setenv(string(semconv.ServiceInstanceIDKey), "instance"))
-				require.NoError(t, os.Setenv(string(semconv.DeploymentEnvironmentKey), "abcd"))
-			},
-			expErr: fmt.Errorf("invalid env: %w", config.ErrInvalidConfig),
-		},
-	}
-
-	for name, tc := range tcs {
-		t.Run(name, func(t *testing.T) {
-			// Given:
-			defer os.Unsetenv(string(semconv.ServiceNameKey))
-			defer os.Unsetenv(string(semconv.ServiceNamespaceKey))
-			defer os.Unsetenv(string(semconv.ServiceVersionKey))
-			defer os.Unsetenv(string(semconv.ServiceInstanceIDKey))
-			defer os.Unsetenv(string(semconv.DeploymentEnvironmentKey))
-
-			tc.givenOSEnvF(t)
-
-			// When:
-			cfg, err := newConfigFromEnv()
-
-			// Then:
-			require.Equal(t, tc.expErr, err)
-			require.EqualValues(t, tc.expCfg, cfg)
 		})
 	}
 }
