+++ apps/golib/app/internal/otel_resource.go
