+++ apps/golib/app/internal/otel_resource_test.go
