+++ apps/golib/app/internal/stubs.go
