+++ apps/golib/app/internal/zap_test.go
@@ -3,29 +3,31 @@ package internal
 import (
 	"testing"
 
-	"github.com/kneadCODE/crazycat/apps/golib/app/config"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/sdk/resource"
+	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
 )
 
 func Test_NewZap(t *testing.T) {
-	// Given: && When:
-	l, err := NewZap(config.Config{Name: "name", Env: config.EnvDev})
+	// Given:
+	res := resource.NewWithAttributes(
+		semconv.SchemaURL,
+		semconv.DeploymentEnvironment("development"),
+	)
+
+	// When:
+	l, err := NewZap(true, res)
 
 	// Then:
 	require.NoError(t, err)
 	require.NotNil(t, l)
+	l.Info("testing")
 
-	// Given: && When:
-	l, err = NewZap(config.Config{Name: "name", Env: config.EnvStaging})
-
-	// Then:
-	require.NoError(t, err)
-	require.NotNil(t, l)
-
-	// Given: && When:
-	l, err = NewZap(config.Config{Name: "name", Env: config.EnvProd})
+	// When:
+	l, err = NewZap(false, res)
 
 	// Then:
 	require.NoError(t, err)
 	require.NotNil(t, l)
+	l.Info("testing")
 }
