+++ apps/golib/app/span_test.go
@@ -1,4 +1,4 @@
-package app2
+package app
 
 import (
 	"context"
